version 0.0.0.9117

+++ R/install_functions.R
@@ -21,7 +21,7 @@ easy_init <- function() {
   # get the packages we need
   packages <- renv::dependencies()$Package
   purrr::walk(.x = packages,
-              .f = link_cache_to_proj)
+              .f = safely_hydrate)
 
   sync_cache()
   write_cache_binary_pkg_catalog()
@@ -346,9 +346,20 @@ link_cache_to_proj <- function(package) {
     cli::cli_alert_success("Linking to {.emph {package}} in the {.emph binary} cache.")
   }
 
+  # create the new link path into the project library
   new_link_path <- file.path(.libPaths()[1], pname)
-  unlink(new_link_path, recursive = T)
-  invisible(file.symlink(to = new_link_path, from = link_path))
+
+  # double check to be sure you don't recursively delete the parent directory which is the project library
+  # this happens if the package name is "" which happens if there is a trailing comma in dependency list
+  # this should also be caught later down where we strip out empty strings from deps
+  if (pname %in% list.dirs(.libPaths()[1L], full.names = FALSE, recursive = FALSE)) {
+    # unlink any old versions of the package existing in the project library
+    unlink(new_link_path, recursive = T)
+    # create the new symlink from teh cached version to the project library
+    invisible(file.symlink(to = new_link_path, from = link_path))
+
+  }
+
 
   # get the list of dependencies for the package you just installed
   deps <- read.dcf(file.path(new_link_path, "DESCRIPTION")) |>
